smart_report/requirements.txt
pandas
google-generativeai
scikit-learn
pillow
streamlit